chatbot/requirements.txt
httpx==0.28.1
protobuf==6.30.0
python-dotenv==1.0.1
python-telegram-bot==21.11.1
